Sentiment Analysis using Support Vector Machine and Random Forest
Abstract-Sentiment analysis, is commonly known as opinion mining, is a vital field in natural language processing (NLP) that claims to find out the sentiment or emotion expressed in a given text. This research paper demonstrates an exhaustive survey of sentiment analysis, focusing on the application of machine learning techniques. A comprehensive parametric literature review has been completed to determine the sentiment analysis using SVM and Random Forest. Additionally, the paper covers preprocessing techniques, feature extraction, model training, evaluation, and challenges encountered in sentiment analysis. The findings of this research contribute to a deeper understanding of sentiment analysis and provide insights into the effectiveness of machine learning approaches in this domain. Based on the results obtained, two machine learning algorithms named as Random Forest and SVMwere evaluated based on their accuracy in a classification task. The Random Forest algorithm achieved an accuracy of 0.78564, while SVM outperformed it slightly with an accuracy of 0.80394. Both Random Forest and SVM have demonstrated their strengths in achieving respectable accuracies in the given classification task. These results suggest that SVM, with its slightly higher accuracy of  0.80394,  may be a more suitable choice when accuracy is the primary concern.  However,  the  basic  configuration need and characteristics of the problem at hand should be considered when choosing the better algorithm with better results